pi6 working until 1e-5

+++ src/rings/mod.rs
@@ -3,7 +3,7 @@ pub mod zomega;
 pub mod zomicron;
 pub mod zzeta;
 
-pub use types::{Int, Float};
+pub use types::{Float, Int};
 pub use zomega::ZOmega;
 pub use zomicron::ZOmicron;
 pub use zzeta::ZZeta;
